Update README.md
The purpose of this project is to develop a File Integrity Checker – a security tool that monitors and verifies the integrity of files by calculating and comparing cryptographic hash values. This is a fundamental concept in cybersecurity used to detect unauthorized changes, tampering, or corruption in file systems. The tool was implemented in Python and executed within a Kali Linux environment, leveraging its built-in terminal, Python interpreter, and file system tools